+++ mobx_codegen/lib/src/template/computed.dart
@@ -6,13 +6,15 @@ class ComputedTemplate {
       required this.computedName,
       required this.type,
       required this.name,
-      this.isPrivate = false});
+      this.isPrivate = false,
+      this.isKeepAlive});
 
   final StoreTemplate storeTemplate;
   final String computedName;
   final String type;
   final String name;
   final bool isPrivate;
+  final bool? isKeepAlive;
 
   @override
   // ignore: prefer_single_quotes
@@ -20,5 +22,5 @@ class ComputedTemplate {
   Computed<$type>? $computedName;
 
   @override
-  $type get $name => ($computedName ??= Computed<$type>(() => super.$name, name: '${storeTemplate.parentTypeName}.$name')).value;""";
+  $type get $name => ($computedName ??= Computed<$type>(() => super.$name, name: '${storeTemplate.parentTypeName}.$name'${isKeepAlive != null ? ', keepAlive: $isKeepAlive' : ''})).value;""";
 }