+++ mobx_codegen/test/generator_usage_test.g.dart
@@ -15,6 +15,14 @@ mixin _$TestStore on _TestStore, Store {
   String get fields => (_$fieldsComputed ??=
           Computed<String>(() => super.fields, name: '_TestStore.fields'))
       .value;
+  Computed<String>? _$fieldsKeepAliveComputed;
+
+  @override
+  String get fieldsKeepAlive => (_$fieldsKeepAliveComputed ??= Computed<String>(
+          () => super.fieldsKeepAlive,
+          name: '_TestStore.fieldsKeepAlive',
+          keepAlive: true))
+      .value;
   Computed<String>? _$batchedItemsComputed;
 
   @override
@@ -178,6 +186,26 @@ mixin _$TestStore on _TestStore, Store {
     });
   }
 
+  late final _$lateFieldAtom =
+      Atom(name: '_TestStore.lateField', context: context);
+
+  @override
+  String get lateField {
+    _$lateFieldAtom.reportRead();
+    return super.lateField;
+  }
+
+  bool _lateFieldIsInitialized = false;
+
+  @override
+  set lateField(String value) {
+    _$lateFieldAtom.reportWrite(
+        value, _lateFieldIsInitialized ? super.lateField : null, () {
+      super.lateField = value;
+      _lateFieldIsInitialized = true;
+    });
+  }
+
   @override
   ObservableFuture<String> future() {
     final _$future = super.future();
@@ -262,7 +290,9 @@ batchItem2: ${batchItem2},
 batchItem3: ${batchItem3},
 batchItem4: ${batchItem4},
 errorField: ${errorField},
+lateField: ${lateField},
 fields: ${fields},
+fieldsKeepAlive: ${fieldsKeepAlive},
 batchedItems: ${batchedItems}
     ''';
   }